refactor: Update routing configuration for authentication and protected routes
