+++ phone-test/src/declarations/auth.d.ts
@@ -1,3 +1,4 @@
+import { FetchResult } from '@apollo/client';
 interface LoginInput {
   username: String;
   password: String;
@@ -12,4 +13,19 @@ interface AuthResponseType {
 interface DeprecatedAuthResponseType {
   access_token: String;
   user: UserType;
+}
+
+export interface AuthContextProps {
+  user: UserType | null;
+  status: Status;
+  accessToken: string | null;
+  refreshToken: string | null;
+  login: ({
+    username,
+    password
+  }: {
+    username: string;
+    password: string;
+  }) => Promise<FetchResult<any>>;
+  logout: () => void;
 }
